Move `create_grid` to `laser_utils`
