+++ lasy/profiles/from_openpmd_profile.py
@@ -3,8 +3,8 @@ from scipy.constants import c
 import openpmd_api as io
 from openpmd_viewer import OpenPMDTimeSeries
 from .from_array_profile import FromArrayProfile
-from lasy.utils.laser_utils import field_to_envelope
-from lasy.utils.openpmd_input import reorder_array, create_grid
+from lasy.utils.laser_utils import field_to_envelope, create_grid
+from lasy.utils.openpmd_input import reorder_array
 
 
 class FromOpenPMDProfile(FromArrayProfile):
